py3: standardize return type from backends
Previously, we were decoding the `fpcalc` output, so we were returning unicode
from _fingerprint_file_fpcalc and bytes from _fingerprint_file_audioread. This
leaves the tool's output as encoded data (which is nice and clean since it
avoids an unnecessary step -- we don't need the data as a string) and jives
with the standard library (where base64-encoded data is in bytes).