HTTP Request timeout
For slow connections, maybe a request can take too much time to complete. In addition, when AcoustID web service is not available (due to maintenance, connection error, etc.), the request takes "infinite" to complete. By passing the "timeout" param to the Session object, the request can finish when it is not completed or go infinite if it is its default value (None). Then, when failing, raises a standard TimeoutError so the developer can handle it without using the requests library