Fixed issue with URL encoding
The AcoustID Web Service offers the `meta` parameter to determine which
kind of information should be included in the response. This parameter
can consist of several keywords which are concatenated using `+`.
Unfortunately, this char is typically encoded as `%2B` by e.g. the
Requests Python library which used in this project. This causes the
parameter to be ignored by the server. To prevent this, `urllib` is used
to preserve the `+`.